forgot lang param, use bigserial for tweet/user ids

+++ ex_feels/priv/repo/migrations/20171215022040_create_table_tweets.exs
@@ -6,7 +6,7 @@ defmodule ExFeels.Repo.Migrations.CreateTableTweets do
       add :text, :string
       add :retweet_count, :integer
       add :lang, :string
-      add :tweet_id, :integer
+      add :tweet_id, :bigserial
       add :favorite_count, :integer
       add :created_at, :string
       add :hashtags, {:array, :string}
